requirements.txt
torch>=1.13.0
transformers>=4.30.0
sentence-transformers>=2.2.2
nltk>=3.7
rouge-score>=0.1.2
tqdm>=4.64.0
scikit-learn>=1.2.0
